Flytter søk interfacet til komponenten for at den kan importeres riktig.
