--- packages/familie-header/src/typer.ts
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import { Adressebeskyttelsegradering } from '@navikt/familie-typer';
-
-export interface ISøkeresultat {
-    adressebeskyttelseGradering?: Adressebeskyttelsegradering;
-    fagsakId?: number; // null betyr at det ikke finnes fagsak på personen
-    harTilgang: boolean;
-    ident: string;
-    ikon: React.ReactNode;
-    navn?: string;
-}
